requirements.txt
typer==0.3.2
colorama==0.4.4
shellingham==1.4.0
pytest==6.2.4
deep-translator==1.10.1
pyyaml==5.4.1
requests==2.30.0
six==1.16.0